refactor: reorganize settings into grouped categories
