+++ lib/features/settings/ui/appearance/nav_label_tile.dart
@@ -1,46 +1,42 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:storii/app/config/nav_targets.dart';
 import 'package:storii/app/init.dart';
-import 'package:storii/app/providers/nav_providers.dart';
 import 'package:storii/app/providers/settings_provider.dart';
 import 'package:storii/shared/widgets/app_bottom_sheet.dart';
 
-class StartupNavTile extends ConsumerWidget {
-  const StartupNavTile({super.key});
+class NavLabelTile extends ConsumerWidget {
+  const NavLabelTile({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final startupNav = ref.watch(startupNavProvider).label;
+    final labelBehavior = ref.watch(navLabelBehaviorProvider);
+    final label = _label(labelBehavior);
 
     return ListTile(
-      leading: const Icon(Icons.home),
       trailing: const Icon(Icons.chevron_right),
-      title: Text(l10n.settingsStartupRouteTitle),
-      subtitle: Text(startupNav),
+      title: Text(l10n.settingsNavLabelBehavior),
+      subtitle: Text(label),
       onTap: () {
         AppBottomSheet.show(
           context,
-          title: l10n.settingsStartupRouteTitle,
+          title: l10n.settingsNavLabelBehavior,
           body: Consumer(
             builder: (context, ref, _) {
-              final navTargets = ref.read(effectiveNavTargetsProvider);
-              final startup = ref.watch(startupNavProvider);
+              final behavior = ref.watch(navLabelBehaviorProvider);
 
-              return RadioGroup<NavTarget>(
-                groupValue: startup,
+              return RadioGroup<NavigationDestinationLabelBehavior>(
+                groupValue: behavior,
                 onChanged: (target) {
                   if (target != null) {
                     ref
                         .read(userSettingsProvider.notifier)
-                        .setStartupNav(target);
+                        .setNavLabelBehavior(target);
                   }
                 },
                 child: Column(
                   children: [
-                    ...navTargets.map(
-                      (nav) =>
-                          RadioListTile(value: nav, title: Text(nav.label)),
+                    ...NavigationDestinationLabelBehavior.values.map(
+                      (b) => RadioListTile(value: b, title: Text(_label(b))),
                     ),
                   ],
                 ),
@@ -50,5 +46,13 @@ class StartupNavTile extends ConsumerWidget {
         );
       },
     );
+  }
+
+  String _label(NavigationDestinationLabelBehavior b) {
+    return switch (b) {
+      .alwaysHide => l10n.alwaysHide,
+      .alwaysShow => l10n.alwaysShow,
+      .onlyShowSelected => l10n.onlyShowSelected,
+    };
   }
 }
