+++ lib/features/settings/ui/settings_header.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class SettingsHeader extends StatelessWidget {
+  final String title;
+
+  const SettingsHeader({super.key, required this.title});
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+
+    return Padding(
+      padding: const .fromLTRB(16, 16, 16, 8),
+      child: Text(
+        title,
+        style: theme.textTheme.titleSmall?.copyWith(
+          color: theme.colorScheme.primary,
+          fontWeight: .bold,
+          letterSpacing: 0.5,
+        ),
+      ),
+    );
+  }
+}
