+++ lib/l10n/app_en.arb
@@ -89,7 +89,7 @@
   "pureBlackSubtitle": "High contrast theme",
   "fontScale": "Font Scale",
   "language": "Language",
-  "configNav": "Configure Navigation",
+  "navBar": "Navbar",
   "appearance": "Appearance",
   "theme": "Theme",
   "reset": "Reset",
